Safari-609.1.11.1/LayoutTests/accessibility/mac/secure-text-field-supports-fast-search.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
</head>
<body id="body">

<input type="password" id="password">

<div id="console"></div>

<script>

    description("This tests that secure text fields still expose the parameterized attributes for fast searching.");

    if (window.accessibilityController) {
        var passwordField = accessibilityController.accessibleElementById("password");
        shouldBe("passwordField.parameterizedAttributeNames().trim()", "'AXUIElementsForSearchPredicate'");
    }

</script>

<script src="../../resources/js-test-post.js"></script>
</body>
</html>
