Safari-609.1.11.1/LayoutTests/accessibility/mac/secure-text-field-supports-fast-search-expected.txt


This tests that secure text fields still expose the parameterized attributes for fast searching.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS passwordField.parameterizedAttributeNames().trim() is 'AXUIElementsForSearchPredicate'
PASS successfullyParsed is true

TEST COMPLETE
